Parse NDJSON output.

Turn json or ndjson output from Stanford CoreNLP into tabular format

Parse the json output from CoreNLP, either text files, or NDJSON written to one or multiple files.

corenlp_parse_ndjson(x, cols_to_keep = c("sentence", "index", "word", "pos",
  "lemma"), destfile = NULL, logfile = NULL, threads = 1L,
  progress = TRUE, verbose = TRUE)

corenlp_parse_json(x, cols_to_keep = c("sentence", "index", "word", "pos",
  "ner"), destfile = NULL, logfile = NULL, progress = TRUE)

Arguments

x
character vector, the JSON string(s) to be parsed

cols_to_keep
columns to keep

destfile
a character string naming the file to write to

logfile
a character string naming the file to an error log to; if provided, json strings will be written to this file if parsing the json string string fails

threads
integer

progress
logical

verbose
logical

Value

A character vector with the target files.

Details

The JSON results of applying the Stanford CoreNLP annotators can be written to a streaming JSON file (ndjson format). corenlp_parse_json will parse a json string to a data.frame. If a destfile is specified, the output will be appended to the file provided. Without a destfile, a data.frame is returned. Strings that cannot be parsed are written to the logfile, if it is defined. If filename is present, the function will process one or more files with the output of Stanford CoreNLP in a NDJSON format. If a destfile has been defined during initialization, results are written/appended to that file. Otherwise, a data.frame is returned.

Examples

ndjson_file <- system.file(package = "bignlp", "extdata", "ndjson", "reuters_1.ndjson")
json_string <- readLines(ndjson_file)
dt <- corenlp_parse_json(json_string, cols = c("word", "pos", "lemma"), progress = FALSE)

dt <- corenlp_parse_ndjson(x = ndjson_file, destfile = NULL, progress = FALSE)

#> ... processing ndjson file: /Users/blaette/Lab/github/bignlp/inst/extdata/ndjson/reuters_1.ndjson

dt <- corenlp_parse_ndjson(x = ndjson_file, destfile = NULL, progress = TRUE)

tsv_file <- tempfile()
dt <- corenlp_parse_ndjson(x = ndjson_file, destfile = tsv_file, progress = FALSE)

#> ... processing ndjson file: /Users/blaette/Lab/github/bignlp/inst/extdata/ndjson/reuters_1.ndjson

y <- data.table::fread(tsv_file)
dt <- corenlp_parse_ndjson(x = ndjson_file, destfile = tsv_file, progress = TRUE)
y <- data.table::fread(tsv_file)

ndjson_dir <- system.file(package = "bignlp", "extdata", "ndjson")
tsv_file <- tempfile()
corenlp_parse_ndjson(x = ndjson_dir, destfile = tsv_file, threads = 1L)

#> ... number of files to process: 4

#> Time difference of 0.226609 secs

y <- data.table::fread(tsv_file)
tsv_dir <- tempdir()
tsv_files <- sprintf("%s/reuters_annotated_%d.tsv", tsv_dir, 1L:4L)
corenlp_parse_ndjson(
  x = list.files(ndjson_dir, full.names = TRUE),
  destfile = tsv_files,
  threads = 1L
  )

#> [1] "/var/folders/r6/1k6mxnbj5077980k11xvr0q40000gn/T//RtmpeVsINx/reuters_annotated_1.tsv"
#> [2] "/var/folders/r6/1k6mxnbj5077980k11xvr0q40000gn/T//RtmpeVsINx/reuters_annotated_2.tsv"
#> [3] "/var/folders/r6/1k6mxnbj5077980k11xvr0q40000gn/T//RtmpeVsINx/reuters_annotated_3.tsv"
#> [4] "/var/folders/r6/1k6mxnbj5077980k11xvr0q40000gn/T//RtmpeVsINx/reuters_annotated_4.tsv"

dt <- data.table::rbindlist(lapply(tsv_files, data.table::fread))