Annotate a string.

Use CoreNLP to annotate strings.

corenlp_annotate(x, destfile = NULL,
  corenlp_dir = getOption("bignlp.corenlp_dir"),
  properties_file = getOption("bignlp.properties_file"), method = "json",
  threads = 1L, progress = TRUE, preclean = TRUE, verbose = TRUE)

Arguments

x
A data.table (required to have the columns 'id' and 'text'), or a character vector with input file(s). Files are assumed to be tsv files with two columns, 'id' and 'text'.

destfile
An output file, if threads > 1, a directory where ndjson files will be stored.

corenlp_dir
The directory where corenlp resides.

properties_file
A properties file to configure annotator.

method
The output generated, either "json" (default), "txt", or "xml".

threads
An integer.

progress
Logical, whether to show progress bar.

preclean
Logical, whether to preprocess string.

verbose
Logical, whether to output messages.

Value

The target files will be returned, so that they can serve as input to corenlp_parse_ndjson.

Details

If argument threads is 1, the tagging result is returned, if destfile is NULL. If threads is higher than 1, destfile should be a directory where tagging results will be stored as NDJSON files.

Examples

library(data.table)
reuters_dt <- data.table(
  text = readLines(system.file(package = "bignlp", "extdata", "txt", "reuters.txt"))
)
reuters_dt[["id"]] <- 1L:nrow(reuters_dt)

destfile <- tempfile()
corenlp_dir <- system.file(package = "cleanNLP", "extdata", "stanford-corenlp-full-2016-10-31")
properties_english_fast <- system.file(
  package = "cleanNLP", "extdata", "StanfordCoreNLP-english-fast.properties"
)
options(java.parameters = "-Xmx4g")
y <- corenlp_annotate(
  x = reuters_dt,
  destfile = NULL,
  properties_file = properties_english_fast,
  corenlp_dir = corenlp_dir,
  threads = 1L,
  progress = FALSE
  )

#> Status of the Java Virtual Machine: 1

#> Java version: 1.8.0_144-b01